+++ CHANGELOG.md
@@ -30,6 +30,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - `Attribute::map()` method added to map all attribute values
 - `Header::has()` method added to check if a header attribute / value exist
 - All part attributes are now accessible via linked attribute
+- Restore a message from string `Message::fromString()`
 
 ### Breaking changes
 - NaN